Basic Terrain Colouring for Chosen Vertical Slice
Begun painting some of the ground textures for the part of the city that we have selected for our vertical slice. The main feature is that the beach is painted as sand and the roads are marked out on the terrain. The textures used are placeholder and can be replaced at a later date. Also, no detail has been added yet and should wait until more of our models are in place.