Created Car Spawn Script
The car spawn script can be added to traffic nodes as desired. They will spawn a car on start. At present the script has a public array for the car prefabs it can choose between whilst spawning, however, this should probably be automated before the script is applied to any nodes in the main scene.