Added Cloud Wall and Border Fog
There is now a giant ring of clouds surrounding the playable area, hiding the regions that we won't be making. This cloud mesh can be refined if we want but it works for now. Additionally, as the player gets too close to the edge of the play area, their screen will start to fade to white.