Built Construction Site & Restructured Scene Hierarchy
Built the new play area of the construction site. This can be used as a mission drop-off location. I restructured the scene hierarchy to tidy some of the clutter and better organize for the future. I also did some general tidying of the map, tweaking a few objects such as grounding  buildings and fences that were levitating.